requirements.txt
# Core dependencies
torch>=1.10.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
scipy>=1.7.0

# Configuration and logging
pyyaml>=5.4.0

# Metrics and evaluation
scikit-learn>=0.24.0

# Optional: Experiment tracking
# tensorboard>=2.8.0
# wandb>=0.12.0

# Optional: ROS bag processing
# pyrosbag>=0.1.0

# Development dependencies (for testing)
pytest>=7.0.0
pytest-cov>=3.0.0
